Restructure + metadata + docs
- moved ert::cache_size to cpu::cache_size
- added GIT_REVISION and GIT_DESCRIBE to version.h
- moved manager/declaration.hpp to manager/manager.hpp
- moved manager/definition.hpp to manager/manager.cpp
- removed some unnecessary includes
- updated kokkos documenation
- updated kokkos sample
