+++ source/timemory/environment.hpp
@@ -25,6 +25,9 @@
 #pragma once
 
 #include "timemory/environment/declaration.hpp"
-#include "timemory/environment/definition.hpp"
 #include "timemory/environment/macros.hpp"
 #include "timemory/environment/types.hpp"
+
+#if !defined(TIMEMORY_USE_ENVIRONMENT_EXTERN)
+#    include "timemory/environment/definition.hpp"
+#endif
