+++ source/timemory/manager/CMakeLists.txt
@@ -11,3 +11,6 @@ build_intermediate_library(
     HEADERS             ${header_files}
     SOURCES             ${source_files}
     PROPERTY_DEPENDS    GLOBAL)
+
+list(REMOVE_ITEM source_files ${CMAKE_CURRENT_SOURCE_DIR}/extern.cpp)
+timemory_install_header_files(${source_files})
