+++ source/tools/kokkos-connector/kp_timemory.cpp
@@ -1,4 +1,10 @@
 
+#if defined(FORCE_HIDDEN_VISIBILITY) && !_MSC_VER
+#    define TIMEMORY_INTERNAL __attribute__((visibility("internal")))
+#    define TIMEMORY_EXTERNAL __attribute__((visibility("default")))
+#    define TIMEMORY_VISIBILITY(...) TIMEMORY_INTERNAL
+#endif
+
 #include "timemory/runtime/configure.hpp"
 #include "timemory/timemory.hpp"
